pyproject.toml
[project]
name = "streaming-project"
version = "0.1.0"
description = "Add your description here"

requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "praw>=7.8.1",
    "textblob>=0.19.0",
    "websocket-client>=1.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
